chore: 회원관리, 신고 엔티티 시퀀스 데이터 타입 변경

+++ matzipback/src/main/java/com/matzip/matzipback/report/command/domain/aggregate/Reason.java
@@ -10,7 +10,7 @@ public class Reason {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int reasonCode;
+    private long reasonCode;
 
     private String reasonName;
